CTCP-6098: Moved sortOrGetDrafts to DraftDepartureService
