+++ app/config/Module.scala
@@ -17,8 +17,8 @@
 package config
 
 import com.google.inject.AbstractModule
-import controllers.actions._
-import services._
+import controllers.actions.*
+import services.*
 
 import java.time.Clock
 
@@ -29,7 +29,6 @@ class Module extends AbstractModule {
     bind(classOf[IdentifierAction]).to(classOf[AuthenticatedIdentifierAction]).asEagerSingleton()
     bind(classOf[PaginationAppConfig]).asEagerSingleton()
     bind(classOf[Clock]).toInstance(Clock.systemDefaultZone())
-    bind(classOf[DraftDepartureService]).to(classOf[MongoDraftDepartureService])
     bind(classOf[ReferenceDataService]).to(classOf[ReferenceDataServiceImpl]).asEagerSingleton()
   }
